Propose a new way of bed leveling calibration
First calculate the two planes (with left and right endstops), then calculate the mean plane. 
Then calculate Z height of the screw that is fixed (at the back of the build plate) and height of points where the left and right screws are placed.
The distance we should move the thumbscrews is the difference between the current height and the height at the fixed point.